cambios para la muestra de los datos en promedios

+++ src/pages/CursosPage.tsx
@@ -4,6 +4,7 @@ import { CursoService } from "../services/CursosApi";
 import { toast } from "react-toastify";
 import Formulario from "../components/Formulario";
 import CursoForm from "../components/CursoForm";
+import Swal from "sweetalert2";
 
 const CursosPage = () => {
     const [datos, setDatos] = useState<Curso[]>([]);
@@ -19,16 +20,30 @@ const CursosPage = () => {
     }
 
     const handleDelete = (id: string) => {
-        CursoService.deleteCurso(id)
-            .then(response => {
-                console.log(response);
-                toast.success('El curso ha sido eliminado con éxito!');
-                CursoService.getCurso().then((data) => setDatos(data));
-            })
-            .catch(error => {
-                console.error(error);
-                toast.error('Hubo un error al eliminar el curso');
-            });
+        Swal.fire({
+            title: 'Eliminar',
+            text: '¿Está seguro que desea borrar esta nota?',
+            icon: 'question',
+            showCancelButton: true,
+            confirmButtonText: 'Borrar',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                CursoService.deleteCurso(id)
+                    .then(response => {
+                        console.log(response);
+                        toast.success('El curso ha sido eliminado con éxito!');
+                    })
+                    .catch(error => {
+                        console.error(error);
+                        toast.error('Hubo un error al eliminar el curso');
+                    });
+                actualizarDatos()
+                Swal.fire('Borrado', 'La nota ha sido eliminada con éxito', 'success');
+            } else {
+                Swal.fire('Cancelado', 'La nota no ha sido eliminada', 'info');
+            }
+        });
+
     };
 
     return (
@@ -110,6 +125,7 @@ const CursosPage = () => {
                 <div className="container mx-auto p-4">
                     <div className="card bg-sky-200 m-5 shadow-lg rounded-lg border-8 border-blue-500">
                         <div className="card-header p-4 text-2xl font-bold text-center uppercase">Nombre: {dato.nombre}</div>
+                        <div className="card-header p-4 text-2xl font-bold text-center uppercase">Estudiantes:</div>
                         {dato.estudiantes.length > 0 ? (
                             <div className="card-body p-4 grid grid-cols-3 gap-4">
                                 {dato.estudiantes.map((estudiante: any, index: number) => (
